+++ src/universal/mutations/PromoteToTeamLeadMutation.js
@@ -1,13 +1,14 @@
 import {commitMutation} from 'react-relay'
-import fromTeamMemberId from 'universal/utils/relay/fromTeamMemberId'
-import toTeamMemberId from 'universal/utils/relay/toTeamMemberId'
 
 graphql`
-  fragment PromoteToTeamLeadMutation_teamMember on PromoteToTeamLeadPayload {
-    oldTeamLead {
+  fragment PromoteToTeamLeadMutation_team on PromoteToTeamLeadPayload {
+    team {
       isLead
     }
-    newTeamLead {
+    oldLeader {
+      isLead
+    }
+    newLeader {
       isLead
     }
   }
@@ -19,21 +20,15 @@ const mutation = graphql`
       error {
         message
       }
-      ...PromoteToTeamLeadMutation_teamMember
+      ...PromoteToTeamLeadMutation_team
     }
   }
 `
 
 const PromoteToTeamLeadMutation = (environment, teamMemberId, onError, onCompleted) => {
-  const {teamId} = fromTeamMemberId(teamMemberId)
-  const myTeamMemberId = toTeamMemberId(teamId, environment.userId)
   return commitMutation(environment, {
     mutation,
     variables: {teamMemberId},
-    optimisticUpdater: (store) => {
-      store.get(myTeamMemberId).setValue(false, 'isLead')
-      store.get(teamMemberId).setValue(true, 'isLead')
-    },
     onCompleted,
     onError
   })