+++ src/universal/types/graphql.d.ts
@@ -136,19 +136,9 @@ export interface IUser {
   inactive: boolean | null
 
   /**
-   * true if the user is a part of the supplied orgId
-   */
-  isBillingLeader: boolean | null
-
-  /**
    * The application-specific name, defaults to nickname
    */
   preferredName: string
-
-  /**
-   * the orgs and roles for this user on each
-   */
-  userOrgs: Array<IUserOrg | null> | null
   archivedTasks: ITaskConnection | null
   archivedTasksCount: number | null
 
@@ -197,7 +187,17 @@ export interface IUser {
   /**
    * get a single organization and the count of users by status
    */
-  organization: IOrganization
+  organization: IOrganization | null
+
+  /**
+   * The connection between a user and an organization
+   */
+  organizationUser: IOrganizationUser | null
+
+  /**
+   * A single user that is connected to a single organization
+   */
+  organizationUsers: Array<IOrganizationUser>
 
   /**
    * Get the list of all organizations a user belongs to
@@ -224,13 +224,6 @@ export interface IUser {
    * all the teams the user is a part of that the viewer can see
    */
   tms: Array<string | null> | null
-}
-
-export interface IIsBillingLeaderOnUserArguments {
-  /**
-   * the org for which you want the users
-   */
-  orgId: string
 }
 
 export interface IArchivedTasksOnUserArguments {
@@ -344,6 +337,13 @@ export interface IOrganizationOnUserArguments {
   orgId: string
 }
 
+export interface IOrganizationUserOnUserArguments {
+  /**
+   * the orgId
+   */
+  orgId: string
+}
+
 export interface ITasksOnUserArguments {
   /**
    * the datetime cursor
@@ -424,30 +424,6 @@ export interface IAuthIdentityType {
    * true if the identity provider is a social provider, false otherwise
    */
   isSocial: boolean | null
-}
-
-/**
- * The user/org M:F join, denormalized on the user/org tables
- */
-export interface IUserOrg {
-  __typename: 'UserOrg'
-
-  /**
-   * The orgId
-   */
-  id: string | null
-
-  /**
-   * role of the user in the org
-   */
-  role: OrgUserRole | null
-}
-
-/**
- * The role of the org user
- */
-export const enum OrgUserRole {
-  billingLeader = 'billingLeader'
 }
 
 /**
@@ -750,7 +726,7 @@ export interface ITeam {
   /**
    * true if the viewer is the team lead, else false
    */
-  isLead: boolean | null
+  isLead: boolean
 
   /**
    * The phase of the meeting, usually matches the facilitator phase, be could be further along
@@ -1533,12 +1509,7 @@ export interface IOrganization {
   /**
    * true if the viewer is the billing leader for the org
    */
-  isBillingLeader: boolean | null
-
-  /**
-   * The billing leader of the organization (or the first, if more than 1)
-   */
-  mainBillingLeader: IUser | null
+  isBillingLeader: boolean
 
   /**
    * The name of the organization
@@ -1596,7 +1567,7 @@ export interface IOrganization {
    * The datetime the organization was last updated
    */
   updatedAt: any | null
-  orgMembers: IOrganizationMemberConnection | null
+  organizationUsers: IOrganizationUserConnection
 
   /**
    * The count of active & inactive users
@@ -1609,7 +1580,7 @@ export interface IOrganization {
   billingLeaders: Array<IUser>
 }
 
-export interface IOrgMembersOnOrganizationArguments {
+export interface IOrganizationUsersOnOrganizationArguments {
   after?: string | null
   first?: number | null
 }
@@ -1639,8 +1610,8 @@ export interface ICreditCard {
 /**
  * A connection to a list of items.
  */
-export interface IOrganizationMemberConnection {
-  __typename: 'OrganizationMemberConnection'
+export interface IOrganizationUserConnection {
+  __typename: 'OrganizationUserConnection'
 
   /**
    * Information to aid in pagination.
@@ -1650,7 +1621,7 @@ export interface IOrganizationMemberConnection {
   /**
    * A list of edges.
    */
-  edges: Array<IOrganizationMemberEdge>
+  edges: Array<IOrganizationUserEdge>
 }
 
 /**
@@ -1683,13 +1654,13 @@ export interface IPageInfo {
 /**
  * An edge in a connection.
  */
-export interface IOrganizationMemberEdge {
-  __typename: 'OrganizationMemberEdge'
+export interface IOrganizationUserEdge {
+  __typename: 'OrganizationUserEdge'
 
   /**
    * The item at the end of the edge
    */
-  node: IOrganizationMember | null
+  node: IOrganizationUser
 
   /**
    * A cursor for use in pagination
@@ -1697,12 +1668,73 @@ export interface IOrganizationMemberEdge {
   cursor: string
 }
 
-export interface IOrganizationMember {
-  __typename: 'OrganizationMember'
-  id: string | null
-  organization: IOrganization | null
-  user: IUser | null
-  isBillingLeader: boolean | null
+/**
+ * organization-specific details about a user
+ */
+export interface IOrganizationUser {
+  __typename: 'OrganizationUser'
+
+  /**
+   * orgId::userId
+   */
+  id: string
+
+  /**
+   * true if the user is paused and the orgs are not being billed, else false
+   */
+  inactive: boolean
+
+  /**
+   * the datetime the user first joined the org
+   */
+  joinedAt: any
+
+  /**
+   * The last moment a billing leader can remove the user from the org & receive a refund
+   */
+  newUserUntil: any
+
+  /**
+   * the datetime the user was removed from the organization if they were a new user
+   */
+  newUserRefundAppliedAt: any | null
+
+  /**
+   * FK
+   */
+  orgId: string
+
+  /**
+   * The user attached to the organization
+   */
+  organization: IOrganization
+
+  /**
+   * if not a member, the datetime the user was removed from the org
+   */
+  removedAt: any | null
+
+  /**
+   * role of the user in the org
+   */
+  role: OrgUserRole | null
+
+  /**
+   * FK
+   */
+  userId: string
+
+  /**
+   * The user attached to the organization
+   */
+  user: IUser
+}
+
+/**
+ * The role of the org user
+ */
+export const enum OrgUserRole {
+  billingLeader = 'billingLeader'
 }
 
 export interface IOrgUserCount {
@@ -2096,7 +2128,7 @@ export interface IInvoiceEdge {
   /**
    * The item at the end of the edge
    */
-  node: IInvoice | null
+  node: IInvoice
   cursor: any | null
 }
 
@@ -2486,7 +2518,7 @@ export interface INotificationEdge {
   /**
    * The item at the end of the edge
    */
-  node: Notification | null
+  node: Notification
   cursor: any | null
 }
 
@@ -2690,11 +2722,6 @@ export interface IMutation {
    * Cancel a pending request for an invitee to join the org
    */
   cancelApproval: ICancelApprovalPayload | null
-
-  /**
-   * Cancel an invitation
-   */
-  cancelTeamInvite: ICancelTeamInvitePayload | null
 
   /**
    * Change the team a task is associated with
@@ -2923,11 +2950,6 @@ export interface IMutation {
    * Request to become the facilitator in a meeting
    */
   requestFacilitator: IRequestFacilitatorPayload | null
-
-  /**
-   * Resend an invitation
-   */
-  resendTeamInvite: IResendTeamInvitePayload | null
 
   /**
    * track an event in segment, like when errors are hit
@@ -3203,13 +3225,6 @@ export interface ICancelApprovalOnMutationArguments {
    * org approval id to cancel
    */
   orgApprovalId: string
-}
-
-export interface ICancelTeamInviteOnMutationArguments {
-  /**
-   * The id of the invitation
-   */
-  invitationId: string
 }
 
 export interface IChangeTaskTeamOnMutationArguments {
@@ -3637,13 +3652,6 @@ export interface IRemoveTeamMemberOnMutationArguments {
 
 export interface IRequestFacilitatorOnMutationArguments {
   teamId: string
-}
-
-export interface IResendTeamInviteOnMutationArguments {
-  /**
-   * The id of the invitation
-   */
-  inviteId: string
 }
 
 export interface ISegmentEventTrackOnMutationArguments {
@@ -4918,27 +4926,6 @@ export interface ICancelApprovalPayload {
   archivedSoftTasks: Array<ITask | null> | null
 }
 
-export interface ICancelTeamInvitePayload {
-  __typename: 'CancelTeamInvitePayload'
-  error: IStandardMutationError | null
-
-  /**
-   * The cancelled invitation
-   */
-  invitation: IInvitation | null
-  removedTeamInviteNotification: INotifyTeamInvite | null
-
-  /**
-   * The soft team members that are no longer tentatively on the team
-   */
-  removedSoftTeamMember: ISoftTeamMember | null
-
-  /**
-   * The tasks that belonged to the soft team member
-   */
-  archivedSoftTasks: Array<ITask | null> | null
-}
-
 export interface IChangeTaskTeamPayload {
   __typename: 'ChangeTaskTeamPayload'
   error: IStandardMutationError | null
@@ -5757,8 +5744,9 @@ export interface IPromoteNewMeetingFacilitatorPayload {
 export interface IPromoteToTeamLeadPayload {
   __typename: 'PromoteToTeamLeadPayload'
   error: IStandardMutationError | null
-  oldTeamLead: ITeamMember | null
-  newTeamLead: ITeamMember | null
+  team: ITeam | null
+  oldLeader: ITeamMember | null
+  newLeader: ITeamMember | null
 }
 
 export interface IRejectOrgApprovalPayload {
@@ -5922,7 +5910,8 @@ export interface IRemoveOrgUserPayload {
   /**
    * The organization member that got removed
    */
-  removedOrgMember: IOrganizationMember | null
+  removedOrgMember: IOrganizationUser | null
+  organizationUserId: string | null
 }
 
 /**
@@ -6035,12 +6024,6 @@ export interface IRequestFacilitatorPayload {
   requestor: ITeamMember | null
 }
 
-export interface IResendTeamInvitePayload {
-  __typename: 'ResendTeamInvitePayload'
-  error: IStandardMutationError | null
-  invitation: IInvitation | null
-}
-
 export interface ISegmentEventTrackOptions {
   teamId?: string | null
   orgId?: string | null
@@ -6063,7 +6046,7 @@ export interface ISetOrgUserRolePayload {
   __typename: 'SetOrgUserRolePayload'
   error: IStandardMutationError | null
   organization: IOrganization | null
-  updatedOrgMember: IOrganizationMember | null
+  updatedOrgMember: IOrganizationUser | null
 }
 
 export interface ISetPhaseFocusPayload {
@@ -6564,7 +6547,7 @@ export interface ISubscription {
   slackChannelAdded: IAddSlackChannelPayload
   slackChannelRemoved: IRemoveSlackChannelPayload
   teamSubscription: TeamSubscriptionPayload
-  teamMemberSubscription: TeanMemberSubscriptionPayload
+  teamMemberSubscription: TeamMemberSubscriptionPayload
 }
 
 export interface IAgendaItemSubscriptionOnSubscriptionArguments {
@@ -6629,9 +6612,7 @@ export type IntegrationSubscriptionPayload = IAddProviderPayload | IRemoveProvid
 export type InvitationSubscriptionPayload =
   | IAcceptTeamInvitePayload
   | IApproveToOrgPayload
-  | ICancelTeamInvitePayload
   | IInviteTeamMembersPayload
-  | IResendTeamInvitePayload
 
 export type NotificationSubscriptionPayload =
   | IAcceptTeamInvitationPayload
@@ -6640,7 +6621,6 @@ export type NotificationSubscriptionPayload =
   | IAddTeamPayload
   | IApproveToOrgPayload
   | ICancelApprovalPayload
-  | ICancelTeamInvitePayload
   | IClearNotificationPayload
   | ICreateTaskPayload
   | IDeleteTaskPayload
@@ -6674,7 +6654,7 @@ export interface ISetOrgUserRoleAddedPayload {
   __typename: 'SetOrgUserRoleAddedPayload'
   error: IStandardMutationError | null
   organization: IOrganization | null
-  updatedOrgMember: IOrganizationMember | null
+  updatedOrgMember: IOrganizationUser | null
 
   /**
    * If promoted, notify them and give them all other admin notifications
@@ -6686,7 +6666,7 @@ export interface ISetOrgUserRoleRemovedPayload {
   __typename: 'SetOrgUserRoleRemovedPayload'
   error: IStandardMutationError | null
   organization: IOrganization | null
-  updatedOrgMember: IOrganizationMember | null
+  updatedOrgMember: IOrganizationUser | null
 
   /**
    * If demoted, notify them and remove all other admin notifications
@@ -6697,7 +6677,6 @@ export interface ISetOrgUserRoleRemovedPayload {
 export type TaskSubscriptionPayload =
   | IAcceptTeamInvitePayload
   | ICancelApprovalPayload
-  | ICancelTeamInvitePayload
   | IChangeTaskTeamPayload
   | ICreateGitHubIssuePayload
   | ICreateTaskPayload
@@ -6731,6 +6710,7 @@ export type TeamSubscriptionPayload =
   | INewMeetingCheckInPayload
   | IPromoteFacilitatorPayload
   | IPromoteNewMeetingFacilitatorPayload
+  | IPromoteToTeamLeadPayload
   | IRequestFacilitatorPayload
   | IRemoveOrgUserPayload
   | IRemoveReflectionPayload
@@ -6780,14 +6760,12 @@ export interface IUpdateDragLocationPayload {
   userId: string
 }
 
-export type TeanMemberSubscriptionPayload =
+export type TeamMemberSubscriptionPayload =
   | IAcceptTeamInvitePayload
   | ICancelApprovalPayload
-  | ICancelTeamInvitePayload
   | IRemoveTeamMemberPayload
   | IInviteTeamMembersPayload
   | IMeetingCheckInPayload
-  | IPromoteToTeamLeadPayload
   | IRejectOrgApprovalPayload
   | IRemoveOrgUserPayload
   | IUpdateUserProfilePayload