+++ src/universal/utils/relay/toOrgMemberId.js
@@ -1,3 +1,0 @@
-const toOrgMemberId = (orgId, userId) => `${userId}::${orgId}`
-
-export default toOrgMemberId